refractored code for better UI design
